allowing to run IniFile throwing exceptions or returning success code optionally. This is the main task of this branch. It is needed e.g. for embedded controller where no exceptions are available.
This version was developped quite fast and shall be tested soon, 
so many features will not persist: 
All in relation with templates and subclassing will be removed again. 

The mechanism introduced to decide whether exceptions are provided 
is through the new preprocessor switch THROW_PREVENTED
If this is set before including inicpp.h, then code with exceptions will not be compiled. 

The following features are likely to persist: 
class IniField/IniFieldBase: 
- added method asUnconditional, 
- added method toString()
(both throwing no exception)
- improved assignment operator to bool. 

in class DecodeResult: 
- added method isOk()
- eliminated method throwIfError() moved to IniFile

in class IniFile: 
available only if THROW_PREVENTED not set: 
- two constructors with parameter stream and file. 
- method decode(stream)
added methods avoiding throwing exceptions: 
- tryDecode and tryLoad 
- reimplemented methods decode and load throwing exceptions 
  based on tryDecode and tryLoad 

The idea is that code throwing exceptions are embraced in 
#ifndef THROW_PREVENTED
#endif 
and that the rest of the code does not depend on these sections. 

In testcases: 
reworked so that the new methods can be used. 
added many of these braces but restructured to avoid too many of them.