For Bugfix 'notify of output stream errors': modified core encoding method changed signature void encode(OutStreamInterface &oStream) const --> DecEncResult tryEncode(OutStreamInterface &oStream)
of course, rename had implication but also remove of const. 
But then also since former DecodeResult now encloses encoding failures, 
it needed to be renamed DecodeResult-->DecEncResult. 
Likewise DecodeErrorCode-->DecEncErrorCode

core change in tryEncode: 
- added test for isOpen 
- added test for bad bit 
- added closing stream (deactivated still)
- added linenumbers to notify about where an error occurred. 
As a consequence, needed return valu. 
Now it is much parallel with tryDecode. 

As a consequence, had to extend DecEncResult and related and OutStreamInterface and related. 
Extension of DecEncResult is just extension of DecEncErrorCode 
and this one triggers extension in method throwIfError() 

DecEncErrorCode: 
- renamed: STREAM_OPEN_FAILED-->STREAM_OPENR_FAILED
- added STREAM_OPENW_FAILED, STREAM_WRITE_FAILED

in method throwIfError 
added and modified branches changed/added in DecEncErrorCode. 

in OutStreamInterface and in OutStream: 
- added method isOpen
- added method bad
- added method close
Now it is much paralle with InStreamInterface and InStream, resp.