replaced class InStream by InFileStream and InStringStream Bugfix notify on encode errors by exception completed (except tests) The core is to transform OutStream into t_OutStream template and to add specialized subclasses OutStringStream and above all OutFileStream overwriting methods isOpen and close. That way it can be detected if a file cannot be opened for write.
The rest of the additions and changes affects methods tryEncode, encode and save 
which are implemented now in parallel to tryDecode, decode and save. 

Parallelity is also in new classes and class templates. 

added more specialized methods tryEncode and encode and save 
base implementations of tryEncode with parameter ostream 
on new classes t_OutSteam, OutStringStream