reimplement map in terms of vector: avoid reordering, 1 step towards mbed-ability The new implementation keeps ordering of key-value pairs, in particular when reading a file and writing it again, possibly with modified values. As map, also vector header require too much memory for application on mbed lpc1768 with ethernet. But we come one step further to memory sparing implementation.
Added template class NamedObj as entries of vector. 
This substitutes for map. 
Added class t_ResMapInterface as a new base class of t_ResMap 
and new class t_ResVMap, 
which has the same interface but is to substiture for T_ResMap. 

Replaced any occurrence of t_ResMap by t_ResVMap 

Commented out definition of class t_ResMap. 

in tests: adapted so that ordering is preserved. 
Also modified test files exampleNl.ini and exampleNoNl.ini 
to avoid keys and section names in alphapetic order 
so that one can see that the input order is preserved.